Updated project files to work with wx event handlers, removed some singletons

+++ src/wx/xrc/xh_aui.cpp
@@ -24,10 +24,11 @@
 wxIMPLEMENT_DYNAMIC_CLASS( wxAuiXmlHandler, wxXmlResourceHandler )
 
 wxAuiXmlHandler::wxAuiXmlHandler()
-                : wxXmlResourceHandler(),
-                  m_manager(NULL),
-                  m_window(NULL),
-                  m_isInside(false)
+:
+wxXmlResourceHandler(),
+m_manager   (NULL),
+m_window    (NULL),
+m_isInside  (false)
 {
     XRC_ADD_STYLE( wxAUI_MGR_ALLOW_ACTIVE_PANE );
     XRC_ADD_STYLE( wxAUI_MGR_ALLOW_FLOATING );
@@ -43,10 +44,10 @@ wxAuiXmlHandler::wxAuiXmlHandler()
 
 wxAuiManager *wxAuiXmlHandler::GetAuiManager( wxWindow *managed )
 {
-    for ( unsigned int i = 0; i < m_managers.GetCount(); i++ )
+    for( size_t i = 0; i < m_managers.GetCount(); i++ )
     {
         wxAuiManager *mgr = m_managers.Item( i );
-        if ( mgr && ( mgr->GetManagedWindow() == managed ) )
+        if( mgr && ( mgr->GetManagedWindow() == managed ) )
             return mgr;
     }
     return NULL;
@@ -55,10 +56,10 @@ wxAuiManager *wxAuiXmlHandler::GetAuiManager( wxWindow *managed )
 void wxAuiXmlHandler::OnManagedWindowClose( wxWindowDestroyEvent &event )
 {
     wxWindow *window = wxDynamicCast( event.GetEventObject(), wxWindow );
-    for ( unsigned int i = 0; i < m_managers.GetCount(); i++ )
+    for( size_t i = 0; i < m_managers.GetCount(); i++ )
     {
         wxAuiManager *mgr = m_managers.Item( i );
-        if ( mgr && ( mgr->GetManagedWindow() == window ) )
+        if( mgr && ( mgr->GetManagedWindow() == window ) )
         {
             mgr->UnInit();
             m_managers.Remove( mgr );
@@ -70,26 +71,26 @@ void wxAuiXmlHandler::OnManagedWindowClose( wxWindowDestroyEvent &event )
 
 wxObject *wxAuiXmlHandler::DoCreateResource()
 {
-    if ( m_class == "wxAuiManager" )
+    if( m_class == "wxAuiManager" )
     {
         wxAuiManager *manager = NULL;
 
         wxXmlNode *windowNode = m_node->GetParent();
-        while ( windowNode )
+        while( windowNode )
         {
-            if ( ( windowNode->GetName() == "object" ||
+            if( ( windowNode->GetName() == "object" ||
                     windowNode->GetName() == "object_ref") )
             {
                 wxString className = "class";
                 wxString name      = "name";
 
-                if ( windowNode->HasAttribute( className ) &&
+                if( windowNode->HasAttribute( className ) &&
                      windowNode->HasAttribute( name ) )
                 {
                     className = windowNode->GetAttribute( className );
                     name      = windowNode->GetAttribute( name );
 
-                    if ( className != "wxAuiPaneInfo" &&
+                    if( className != "wxAuiPaneInfo" &&
                          className != "wxStatusBar"   &&
                          className != "wxToolBar"     &&
                          className != "wxMenuBar" )
@@ -97,7 +98,7 @@ wxObject *wxAuiXmlHandler::DoCreateResource()
                         // Using XRCID here causes undefined references when linking
                         wxWindow *window = wxWindow::FindWindowByName( name );
 
-                        if ( window )
+                        if( window )
                         {
                             bool          old_ins = m_isInside;
                             wxAuiManager *old_mgr = m_manager;
@@ -117,7 +118,7 @@ wxObject *wxAuiXmlHandler::DoCreateResource()
 
                             CreateChildren( manager );
 
-                            if ( HasParam("perspective") )
+                            if( HasParam("perspective") )
                                 manager->LoadPerspective( GetParamValue("perspective") );
 
                             m_window   = old_win;
@@ -132,15 +133,15 @@ wxObject *wxAuiXmlHandler::DoCreateResource()
 
         return manager;
     }
-    else if ( m_class == "wxAuiPaneInfo" )
+    else if( m_class == "wxAuiPaneInfo" )
     {
         wxXmlNode *node   = GetParamNode("object");
         wxWindow  *window = NULL;
 
-        if ( !node )
+        if( !node )
             node = GetParamNode("object_ref");
 
-        if ( node )
+        if( node )
         {
             bool old_ins = m_isInside;
             m_isInside = false;
@@ -150,63 +151,63 @@ wxObject *wxAuiXmlHandler::DoCreateResource()
             m_isInside = old_ins;
             window = wxDynamicCast( object, wxWindow );
 
-            if ( window == NULL && object != NULL )
+            if( window == NULL && object != NULL )
             {
                 ReportError( node, _("wxAuiPaneInfo child must be a window.") );
             }
         }
 
-        if ( window )
+        if( window )
         {
             wxAuiPaneInfo paneInfo = wxAuiPaneInfo();
 
-            wxString name = GetName();          paneInfo.Name( name );
+            wxString name = GetName();         paneInfo.Name( name );
 // Caption
-            if ( HasParam("caption") )          paneInfo.Caption( GetText("caption") );
-            if ( HasParam("caption_visible") )  paneInfo.CaptionVisible( GetBool("caption_visible") );
+            if( HasParam("caption") )          paneInfo.Caption( GetText("caption") );
+            if( HasParam("caption_visible") )  paneInfo.CaptionVisible( GetBool("caption_visible") );
 // Buttons
-            if ( HasParam("close_button") )     paneInfo.CloseButton( GetBool("close_button") );
-            if ( HasParam("minimize_button") )  paneInfo.MinimizeButton( GetBool("minimize_button") );
-            if ( HasParam("maximize_button") )  paneInfo.MaximizeButton( GetBool("maximize_button") );
-            if ( HasParam("pin_button") )       paneInfo.PinButton( GetBool("pin_button") );
-            if ( HasParam("gripper") )          paneInfo.Gripper( GetBool("gripper") );
+            if( HasParam("close_button") )     paneInfo.CloseButton( GetBool("close_button") );
+            if( HasParam("minimize_button") )  paneInfo.MinimizeButton( GetBool("minimize_button") );
+            if( HasParam("maximize_button") )  paneInfo.MaximizeButton( GetBool("maximize_button") );
+            if( HasParam("pin_button") )       paneInfo.PinButton( GetBool("pin_button") );
+            if( HasParam("gripper") )          paneInfo.Gripper( GetBool("gripper") );
 // Appearance
-            if ( HasParam("pane_border") )      paneInfo.PaneBorder( GetBool("pane_border") );
+            if( HasParam("pane_border") )      paneInfo.PaneBorder( GetBool("pane_border") );
 // State
-            if ( HasParam("dock") )             paneInfo.Dock();
-            else if ( HasParam("float") )       paneInfo.Float();
+            if( HasParam("dock") )             paneInfo.Dock();
+            else if( HasParam("float") )       paneInfo.Float();
 
 // Dockable Directions
-            if ( HasParam("top_dockable") )     paneInfo.TopDockable( GetBool("top_dockable") );
-            if ( HasParam("bottom_dockable") )  paneInfo.BottomDockable( GetBool("bottom_dockable") );
-            if ( HasParam("left_dockable") )    paneInfo.LeftDockable( GetBool("left_dockable") );
-            if ( HasParam("right_dockable") )   paneInfo.RightDockable( GetBool("right_dockable") );
+            if( HasParam("top_dockable") )     paneInfo.TopDockable( GetBool("top_dockable") );
+            if( HasParam("bottom_dockable") )  paneInfo.BottomDockable( GetBool("bottom_dockable") );
+            if( HasParam("left_dockable") )    paneInfo.LeftDockable( GetBool("left_dockable") );
+            if( HasParam("right_dockable") )   paneInfo.RightDockable( GetBool("right_dockable") );
 // Behaviours
-            if ( HasParam("dock_fixed") )       paneInfo.DockFixed( GetBool("dock_fixed") );
-            if ( HasParam("resizable") )        paneInfo.Resizable( GetBool("resizable") );
-            if ( HasParam("movable") )          paneInfo.Movable( GetBool("movable") );
-            if ( HasParam("floatable") )        paneInfo.Floatable( GetBool("floatable") );
+            if( HasParam("dock_fixed") )       paneInfo.DockFixed( GetBool("dock_fixed") );
+            if( HasParam("resizable") )        paneInfo.Resizable( GetBool("resizable") );
+            if( HasParam("movable") )          paneInfo.Movable( GetBool("movable") );
+            if( HasParam("floatable") )        paneInfo.Floatable( GetBool("floatable") );
 // Sizes
-            if ( HasParam("floating_size") )    paneInfo.FloatingSize( GetSize("floating_size") );
-            if ( HasParam("min_size") )         paneInfo.MinSize( GetSize("min_size") );
-            if ( HasParam("max_size") )         paneInfo.MaxSize( GetSize("max_size") );
-            if ( HasParam("best_size") )        paneInfo.BestSize( GetSize("best_size") );
+            if( HasParam("floating_size") )    paneInfo.FloatingSize( GetSize("floating_size") );
+            if( HasParam("min_size") )         paneInfo.MinSize( GetSize("min_size") );
+            if( HasParam("max_size") )         paneInfo.MaxSize( GetSize("max_size") );
+            if( HasParam("best_size") )        paneInfo.BestSize( GetSize("best_size") );
 // Positions
-            if ( HasParam("row") )              paneInfo.Row( GetLong("row") );
-            if ( HasParam("layer") )            paneInfo.Layer( GetLong("layer") );
-            if ( HasParam("default_pane") )     paneInfo.DefaultPane();
-            else if( HasParam("toolbar_pane") ) paneInfo.ToolbarPane();
+            if( HasParam("row") )              paneInfo.Row( GetLong("row") );
+            if( HasParam("layer") )            paneInfo.Layer( GetLong("layer") );
+            if( HasParam("default_pane") )     paneInfo.DefaultPane();
+            else if( HasParam("toolbar_pane") )paneInfo.ToolbarPane();
 
 // Directions - CenterPane()/CentrePane != Center()/Centre()
-            if ( HasParam("center_pane") ||
-                 HasParam("centre_pane") )      paneInfo.CenterPane();
-            if ( HasParam("direction") )        paneInfo.Direction( GetLong("direction") );
-            else if ( HasParam("top") )         paneInfo.Top();
-            else if ( HasParam("bottom") )      paneInfo.Bottom();
-            else if ( HasParam("left") )        paneInfo.Left();
-            else if ( HasParam("right") )       paneInfo.Right();
-            else if ( HasParam("center") ||
-                      HasParam("centre") )      paneInfo.Center();
+            if( HasParam("center_pane") ||
+                 HasParam("centre_pane") )     paneInfo.CenterPane();
+            if( HasParam("direction") )        paneInfo.Direction( GetLong("direction") );
+            else if( HasParam("top") )         paneInfo.Top();
+            else if( HasParam("bottom") )      paneInfo.Bottom();
+            else if( HasParam("left") )        paneInfo.Left();
+            else if( HasParam("right") )       paneInfo.Right();
+            else if( HasParam("center") ||
+                     HasParam("centre") )      paneInfo.Center();
 
             m_manager->AddPane( window, paneInfo );
             m_manager->Update();
@@ -222,7 +223,7 @@ wxObject *wxAuiXmlHandler::DoCreateResource()
 
 bool wxAuiXmlHandler::CanHandle( wxXmlNode *node )
 {
-    return (!m_isInside && IsOfClass( node, "wxAuiManager")) ||
+    return( !m_isInside && IsOfClass( node, "wxAuiManager") )||
             (m_isInside && IsOfClass( node, "wxAuiPaneInfo"));
 }
